# schemas/to-be/api/openapi.yaml
openapi: 3.0.3
info:
  title: WarmHome API
  version: 1.0.0
  description: |
    ## Допущение по аутентификации и авторизации

    Все внешние REST-запросы проходят через API Gateway, который:
    - валидирует заголовок `Authorization: Bearer <JWT>` (токен выдан внешним Identity Provider);
    - выполняет проверку прав доступа пользователя к доменным ресурсам (например, доступ к `houseId`).

    Downstream-микросервисы:
    - не получают и не используют `userId`;
    - не выполняют повторную валидацию JWT и не принимают решения об авторизации;
    - работают только с доменными идентификаторами (`houseId`, `deviceId` и т.п.), считая запрос доверенным после проверки на Gateway.

    ## Допущение по моделям устройств

    В системе существует предварительно заполненный каталог моделей устройств (Device Model Catalog),
    управляемый администратором.
    При добавлении устройства пользователь выбирает модель из каталога, после чего `deviceModelId`
    передаётся в запросе `POST /api/v1/devices` как ссылка на справочную сущность.
servers:
  - url: https://api.warmhome.example.com
...
paths:
...
  /api/v1/devices/{deviceId}/commands:
    post:
      tags: [Commands]
      summary: Отправка команды устройству
      description: |
        Создаёт новую команду управления устройством.
        Команда обрабатывается асинхронно и доставляется устройству через адаптеры.
      parameters:
        - name: deviceId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCommandRequest" }
            examples:
              turnOn:
                value:
                  commandType: "HEAT_RELAY_SET"
                  payload:
                    enabled: true
      responses:
        "202":
          description: Команда принята к обработке
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Command" }
              examples:
                accepted:
                  value:
                    id: "aa0c9e7d-0e7b-4c9d-9a2c-8e0a1d2b3c4d"
                    externalDeviceId: "0cbf1a0a-9b83-4a4f-9b21-112233445566"
                    commandType: "HEAT_RELAY_SET"
                    payload: { enabled: true }
                    status: "pending"
                    createdAt: "2026-01-11T10:01:00Z"
        "400":
          description: Ошибка валидации запроса
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                validation:
                  value:
                    code: "VALIDATION_ERROR"
                    message: "Invalid request"
                    details:
                      field: "commandType"
                      reason: "required"
        "404":
          description: Устройство не найдено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                notFound:
                  value:
                    code: "NOT_FOUND"
                    message: "Device not found"
                    details:
                      resource: "device"
                      id: "0cbf1a0a-9b83-4a4f-9b21-112233445566"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                forbidden:
                  value:
                    code: "FORBIDDEN"
                    message: "Access denied"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/devices/{deviceId}/telemetry:
    get:
      tags: [Telemetry]
      summary: Получение телеметрии устройства
      description: |
        Возвращает телеметрию устройства в виде временного ряда.
        Может использоваться как для получения истории измерений,
        так и для получения текущего значения метрики (например, через `limit=1`).
      parameters:
        - name: deviceId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: metric
          in: query
          required: false
          schema: { type: string, example: "temperature" }
        - name: from
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Телеметрия успешно получена
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/TelemetryData" }
              examples:
                ok:
                  value:
                    items:
                      - id: "f1e2d3c4-b5a6-4f00-9a99-123456789000"
                        externalDeviceId: "0cbf1a0a-9b83-4a4f-9b21-112233445566"
                        metric: "temperature"
                        value: "22.6"
                        recordedAt: "2026-01-11T09:59:00Z"
                      - id: "0aa1bb2c-3dd4-4ee5-9ff0-098765432111"
                        externalDeviceId: "0cbf1a0a-9b83-4a4f-9b21-112233445566"
                        metric: "temperature"
                        value: "22.7"
                        recordedAt: "2026-01-11T10:00:00Z"
        "404":
          description: Устройство не найдено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                notFound:
                  value:
                    code: "NOT_FOUND"
                    message: "Device not found"
                    details:
                      resource: "device"
                      id: "0cbf1a0a-9b83-4a4f-9b21-112233445566"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                forbidden:
                  value:
                    code: "FORBIDDEN"
                    message: "Access denied"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            internal:
              value:
                code: "INTERNAL_ERROR"
                message: "Unexpected error"
  schemas:
...
    CreateCommandRequest:
      type: object
      required: [commandType]
      properties:
        commandType: { type: string, example: "HEAT_RELAY_SET" }
        payload:
          type: object
          nullable: true
          additionalProperties: true
    Command:
      type: object
      required: [id, externalDeviceId, commandType, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        externalDeviceId: { type: string, format: uuid }
        commandType: { type: string }
        payload:
          type: object
          nullable: true
          additionalProperties: true
        status: { type: string, enum: [pending, succeeded, failed] }
        createdAt: { type: string, format: date-time }
    TelemetryData:
      type: object
      required: [id, externalDeviceId, metric, value, recordedAt]
      properties:
        id: { type: string, format: uuid }
        externalDeviceId: { type: string, format: uuid }
        metric: { type: string, example: "temperature" }
        value: { type: string, example: "22.7" }
        recordedAt: { type: string, format: date-time }
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, NOT_FOUND, CONFLICT, FORBIDDEN, INTERNAL_ERROR]
        message:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true